requirements.txt
pytesseract
Pillow
PyQt6
